refactor: 연산부분 유효성 검증 메서드의 클래스 이동
