+++ src/main/java/calculator/model/Arithmetic.java
@@ -1,6 +1,7 @@
 package calculator.model;
 
-import java.util.ArrayList;
+import calculator.util.ErrorMessage;
+
 import java.util.Arrays;
 import java.util.List;
 
@@ -8,7 +9,23 @@ public class Arithmetic {
     private List<Integer> arithmetic;
 
     public Arithmetic(Delimiter delimiter, InputData inputData) {
+        if(inputData.isDelimiter()){
+            validateWithCustomDelimiter(inputData);
+        }
+        if(!inputData.isDelimiter()){
+            validateWithDefaultDelimiter(inputData);
+        }
         this.arithmetic = splitArithmetic(delimiter, inputData.convertCalculatorPart());
+    }
+
+    public void validateWithDefaultDelimiter(InputData inputData){
+        if (!inputData.convertCalculatorPart().matches("^[0-9,:]*$")) {
+            throw new IllegalArgumentException(ErrorMessage.INPUT_DATA_WITH_DEFAULT_DELIMITER_FORMAT.getError());
+        }
+    }
+
+    public void validateWithCustomDelimiter(InputData inputData){
+
     }
 
     public List<Integer> splitArithmetic(Delimiter delimiter, String calculatorPart) {
