+++ src/test/java/calculator/model/InputDataTest.java
@@ -38,22 +38,4 @@ class InputDataTest {
         InputData inputData = new InputData("//;.!\\n1;2.3!4");
         assertThat(inputData.convertCalculatorPart()).isEqualTo("1;2.3!4");
     }
-
-    @ParameterizedTest
-    @ValueSource(strings = {"1,2,3!4","a1,2:3","1,가,2,3"})
-    @DisplayName("기본 구분자를 사용할 때 숫자와 기본구분자 이외의 문자가 있으면 예외가 발생한다")
-    void isInvalidCalculatorWithDefaultDelimiter(){
-        assertThatThrownBy(()->new InputData("1,2,3!4"))
-                .isInstanceOf(IllegalArgumentException.class)
-                .hasMessage(ErrorMessage.INPUT_DATA_WITH_DEFAULT_DELIMITER_FORMAT.getError());
-    }
-
-    @ParameterizedTest
-    @ValueSource(strings = {"1,2,3,4,",":1:2:3","1:2,3"})
-    @DisplayName("기본 구분자를 사용할 때 숫자와 기본구분자만 있으면 예외가 발생하지 않는다")
-    void isValidCalculatorWithDefaultDelimiter(String input){
-        assertThatCode(()->new InputData(input))
-                .doesNotThrowAnyException();
-    }
-
 }
